# plugins/cmd/subnet-plugin/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Simple wrapper around Aruba Cloud API to provide consistency of API response for Krateo Operator Generator (KOG)",
        "title": "Aruba Cloud Subnet Plugin API for Krateo Operator Generator (KOG)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Krateo Support",
            "url": "https://krateo.io",
            "email": "contact@krateoplatformops.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/projects/{projectId}/providers/Aruba.Network/vpcs/{vpcId}/subnets": {
            "get": {
                "description": "List Subnets on Aruba Cloud using the provided project and vpc details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List Subnets on Aruba Cloud",
                "operationId": "list-subnets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "VPC ID",
                        "name": "vpcId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API version (e.g., 1.0)",
                        "name": "api-version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter expression",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort expression",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Projection expression",
                        "name": "projection",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit for pagination",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token (Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of subnets",
                        "schema": {
                            "$ref": "#/definitions/cmd_subnet-plugin_handlers.FlattenedSubnetListResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create a new Subnet on Aruba Cloud using the provided project and vpc details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new Subnet on Aruba Cloud",
                "operationId": "post-subnet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "VPC ID",
                        "name": "vpcId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API version (e.g., 1.0)",
                        "name": "api-version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token (Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Subnet creation request body",
                        "name": "subnetCreate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_subnet-plugin_handlers.FlattenedCreateSubnetRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Subnet details",
                        "schema": {
                            "$ref": "#/definitions/cmd_subnet-plugin_handlers.FlattenedSubnetResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/projects/{projectId}/providers/Aruba.Network/vpcs/{vpcId}/subnets/{id}": {
            "get": {
                "description": "Get a Subnet from Aruba Cloud using the provided project, vpc, and subnet details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a Subnet from Aruba Cloud",
                "operationId": "get-subnet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "VPC ID",
                        "name": "vpcId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subnet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API version (e.g., 1.0)",
                        "name": "api-version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "if the resource exists in status 'Deleted', returns NotFound according to the value of this flag",
                        "name": "ignoreDeletedStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token (Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subnet details",
                        "schema": {
                            "$ref": "#/definitions/cmd_subnet-plugin_handlers.FlattenedSubnetResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update a Subnet on Aruba Cloud using the provided project, vpc, and subnet details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a Subnet on Aruba Cloud",
                "operationId": "put-subnet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "VPC ID",
                        "name": "vpcId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subnet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API version (e.g., 1.0)",
                        "name": "api-version",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer Token (Bearer \u003ctoken\u003e)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Subnet update request body",
                        "name": "subnetUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_subnet-plugin_handlers.FlattenedUpdateSubnetRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subnet details",
                        "schema": {
                            "$ref": "#/definitions/cmd_subnet-plugin_handlers.FlattenedSubnetResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "cmd_subnet-plugin_handlers.FlattenedCreateSubnetRequestDto": {
            "type": "object",
            "properties": {
                "location": {
                    "$ref": "#/definitions/cmd_subnet-plugin_handlers.LocationDto"
                },
                "name": {
                    "type": "string"
                },
                "properties": {
                    "$ref": "#/definitions/cmd_subnet-plugin_handlers.SubnetPropertiesDto"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "cmd_subnet-plugin_handlers.FlattenedSubnetListResponseDto": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                },
                "self": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd_subnet-plugin_handlers.FlattenedSubnetResponseDto"
                    }
                }
            }
        },
...
        "cmd_subnet-plugin_handlers.FlattenedUpdateSubnetRequestDto": {
            "type": "object",
            "properties": {
                "location": {
                    "$ref": "#/definitions/cmd_subnet-plugin_handlers.LocationDto"
                },
                "name": {
                    "type": "string"
                },
                "properties": {
                    "$ref": "#/definitions/cmd_subnet-plugin_handlers.SubnetUpdatePropertiesDto"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "cmd_subnet-plugin_handlers.LocationDto": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "string"
                }
            }
        },
...
        "cmd_subnet-plugin_handlers.SubnetUpdatePropertiesDto": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "boolean"
                }
            }
        },
...
    }
}